read query data based on entity type
